requirements.txt
fastapi
uvicorn[standard]
mlflow
scikit-learn
pandas
numpy
joblib
pydantic
python-dotenv
requests
loguru
google-cloud-storage
google-cloud-artifact-registry
tqdm
pyyaml
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace